telegram_legacy_v1/requirements.txt
python-telegram-bot==20.*
requests
python-dotenv
appdirs
